Close the second acceptance pass
The second reviewer returned 'not accepted' again, with eight findings
and nine mutations the suite let through. All are addressed.

Ordering was judged only on rows that survived validation, so an
out-of-order benchmark row with a bad price was dropped and the fixture
then looked well ordered - the spec says any benchmark row on or before
a training row condemns the file. Order is now judged on every row
whose date and split can be read at all. Equality on the boundary was
also accepted; the requirement is strictly later, and both cases now
have tests.

Identity was only defended against different source references. Two
rows sharing a reference, date and SKU but differing elsewhere were
both published under one scenario_id, and a duplicated row went out
twice - a doubled fact in the training statistics. Any repeat of an
identity within a run is now refused, while a repeated run still hands
off again, because that is at-least-once and downstream deduplicates.

Failure handling conflated three things the guide separates. A missing
schema became one rejection per row rather than a deployment failure,
and a publisher error was filed as if the source data were at fault -
the guide forbids reporting such a cycle as successful. Publication now
sits outside the per-row catch, and a missing schema throws.

Task 4.3 was ticked for a structured type with nothing logging it;
cycle start and finish and every rejection are now logged, via
java.util.logging rather than a new dependency for six lines of output.

Two defects in the preparation script: cells were read in document
order rather than by their column letter, so a sparse row shifted every
column after the gap, and the header was taken as the first non-empty
row, which is the workbook title. Both fixed; the script still
reproduces the committed fixture byte for byte.

Also: a header-only fixture was accepted as a clean empty batch; the
stock-level derivation could overflow Int; and the stale 250/0.09
figures survived in a comment.

My own trigger test was vacuous - it compared two outputs that would
match if both triggers did nothing, and a no-op mutation proved it. It
now requires the full set as well as the same set.

93 tests, up from 86.

+++ src/main/kotlin/club/podlodka/snowball/domain/BaselineRecord.kt
@@ -56,5 +56,5 @@ data class BaselineRecord(
      * survives the tuning it allows.
      */
     val stockLevel: StockLevel
-        get() = if (stock >= 2 * baselineSales) StockLevel.HIGH else StockLevel.NORMAL
+        get() = if (stock.toLong() >= 2L * baselineSales) StockLevel.HIGH else StockLevel.NORMAL
 }
